Moved the JS specific call linker to x2cpg
Can now be re-used in js2cpg and CS (will do separate PRs).
