+++ joern-cli/frontends/x2cpg/src/main/scala/io/joern/x2cpg/layers/CallGraph.scala
@@ -3,7 +3,10 @@ package io.joern.x2cpg.layers
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.passes.CpgPassBase
 import io.shiftleft.semanticcpg.layers.{LayerCreator, LayerCreatorContext, LayerCreatorOptions}
+import io.shiftleft.semanticcpg.language._
 import io.joern.x2cpg.passes.callgraph.{DynamicCallLinker, MethodRefLinker, StaticCallLinker}
+import io.joern.x2cpg.passes.frontend.JavascriptCallLinker
+import io.shiftleft.codepropertygraph.generated.Languages
 
 object CallGraph {
   val overlayName: String = "callgraph"
@@ -11,7 +14,13 @@ object CallGraph {
   def defaultOpts         = new LayerCreatorOptions()
 
   def passes(cpg: Cpg): Iterator[CpgPassBase] = {
-    Iterator(new MethodRefLinker(cpg), new StaticCallLinker(cpg), new DynamicCallLinker(cpg))
+    val languageSpecificPasses = cpg.metaData.language.lastOption match {
+      case Some(Languages.JSSRC)      => Iterator[CpgPassBase](new JavascriptCallLinker(cpg))
+      case Some(Languages.JAVASCRIPT) => Iterator[CpgPassBase](new JavascriptCallLinker(cpg))
+      case _                          => Iterator[CpgPassBase]()
+    }
+    languageSpecificPasses ++
+      Iterator(new MethodRefLinker(cpg), new StaticCallLinker(cpg), new DynamicCallLinker(cpg))
   }
 
 }
